Accomplished: Added fetchRequests and matches view.
For next time:
Added fetchRequests from requests coming in to the user. Maybe add a notification page for incoming matches and upcoming flights (really ambitious). For that probably add a timestamp object which is the last time the user logged on or keep them there until they either are rejected or denied.